Move cmd/apko/main.go to repo root
Move cmd/cli to pkg/cli

Update as many references to cmd/apko as I can find.
